+++ crates/warp_tui/src/root_view_tests.rs
@@ -4,7 +4,7 @@ use warpui::{AddWindowOptions, UpdateModel};
 use warpui_core::{App, TuiView as _, WindowId};
 
 use super::RootTuiView;
-use crate::sessions::TuiSessions;
+use crate::session_registry::TuiSessions;
 use crate::test_fixtures::{add_test_semantic_selection, add_test_terminal_session};
 
 fn add_root(app: &mut App) -> (WindowId, warpui_core::ViewHandle<RootTuiView>) {
@@ -32,16 +32,17 @@ fn root_projects_only_the_focused_retained_session_view() {
         });
 
         let (first, first_manager) = add_test_terminal_session(&mut app, window_id);
-        let (second, second_manager) = add_test_terminal_session(&mut app, window_id);
         let first_view_id = first.id();
-        let second_view_id = second.id();
         let first_id = app.update_model(&sessions, |sessions, ctx| {
             sessions.add_session(first, first_manager, true, ctx)
         });
         app.read(|ctx| {
             assert_eq!(root.as_ref(ctx).child_view_ids(ctx), vec![first_view_id]);
+            assert!(ctx.check_view_or_child_focused(window_id, &first_view_id));
         });
         let focused_window_view = app.read(|ctx| ctx.focused_view_id(window_id));
+        let (second, second_manager) = add_test_terminal_session(&mut app, window_id);
+        let second_view_id = second.id();
 
         let second_id = app.update_model(&sessions, |sessions, ctx| {
             sessions.add_session(second, second_manager, false, ctx)
@@ -56,12 +57,16 @@ fn root_projects_only_the_focused_retained_session_view() {
         });
         app.read(|ctx| {
             assert_eq!(root.as_ref(ctx).child_view_ids(ctx), vec![second_view_id]);
+            assert!(ctx.check_view_or_child_focused(window_id, &second_view_id));
+            assert_ne!(ctx.focused_view_id(window_id), focused_window_view);
         });
         app.update_model(&sessions, |sessions, ctx| {
             sessions.focus_session(first_id, ctx);
         });
         app.read(|ctx| {
             assert_eq!(root.as_ref(ctx).child_view_ids(ctx), vec![first_view_id]);
+            assert!(ctx.check_view_or_child_focused(window_id, &first_view_id));
+            assert_eq!(ctx.focused_view_id(window_id), focused_window_view);
         });
     });
 }
